backend/requirements.txt
click
colorama
Flask
Flask-SQLAlchemy
Flask-Cors 
itsdangerous
Jinja2
MarkupSafe
SQLAlchemy
Werkzeug
numpy
pandas
python-dateutil
pytz
scikit-learn
scipy
joblib
python-docx
fpdf
greenlet
reportlab
gunicorn   
